requirements.txt
requests>=2.31.0
mss>=9.0.1
pillow>=10.0.0
psutil>=5.9.0
opencv-python>=4.8.0
numpy>=1.24.0
pywin32>=306
python-dotenv>=1.0.0
cryptography>=41.0.0
